Return its position
Add the ability for the card to return its position (in which object it is located) using the Card.getDestination () function.  Implementing dragging and dropping a card into an empty pile.  Implementing dragging and dropping maps from the star to the heaps. Implement dragging cards in star position. The reason for the not actual redrawing of the stars has been found.